Bug Fix for Puzzle (temp)
Temp fix for issue relating to the GetCurrentSelectedPiece() in PieceStorage.cs
Issue was I commented out a line of code to return pieces to the starting position if there was no valid move. However, the GetCurrentSelectedPiece method checks to see if the piece is moved as a parameter for if that piece is the current selected piece.